+++ tests/pattoo_shared_/installation/test_configure.py
@@ -16,7 +16,7 @@ EXEC_DIR = os.path.dirname(os.path.realpath(__file__))
 ROOT_DIR = os.path.abspath(os.path.join(os.path.abspath(os.path.join(
     os.path.abspath(os.path.join(EXEC_DIR, os.pardir)), os.pardir)), os.pardir))
 _EXPECTED = '''\
-{0}pattoo-shared{0}tests{0}test_pattoo_shared{0}installation'''.format(os.sep)
+{0}pattoo-shared{0}tests{0}pattoo_shared_{0}installation'''.format(os.sep)
 if EXEC_DIR.endswith(_EXPECTED) is True:
     # We need to prepend the path in case PattooShared has been installed
     # elsewhere on the system using PIP. This could corrupt expected results
