+++ tests/pattoo_shared_/test_agent.py
@@ -12,7 +12,7 @@ import multiprocessing
 EXEC_DIR = os.path.dirname(os.path.realpath(__file__))
 ROOT_DIR = os.path.abspath(os.path.join(
     os.path.abspath(os.path.join(EXEC_DIR, os.pardir)), os.pardir))
-_EXPECTED = '{0}pattoo-shared{0}tests{0}test_pattoo_shared'.format(os.sep)
+_EXPECTED = '{0}pattoo-shared{0}tests{0}pattoo_shared_'.format(os.sep)
 if EXEC_DIR.endswith(_EXPECTED) is True:
     # We need to prepend the path in case PattooShared has been installed
     # elsewhere on the system using PIP. This could corrupt expected results
@@ -30,7 +30,7 @@ from tests.libraries.configuration import UnittestConfig
 from tests.libraries import general
 
 # Test Daemon importation
-from tests.test_pattoo_shared import test_daemon
+from tests.pattoo_shared_ import test_daemon
 
 
 class TestAgent(unittest.TestCase):
